% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_snv}
\alias{data_snv}
\title{Single nucleotide variant repeat data}
\format{
A dataframe in tidy format with columns for sample, replicate,
variant and variant frequency. This dataset includes different numbers of
variants per sample, and different numbers of replicates per sample. Some
variants are different and some are shared between samples.
}
\source{
Frequency results are randomly generated with R.The code for generating
the data is saved in the data-raw folder.
}
\usage{
data_snv
}
\description{
Single nucleotide variant repeat data
}
\examples{
data_snv
}
\keyword{datasets}
