% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-stats.R
\name{group_stats}
\alias{group_stats}
\title{Calculate statistical variation within groups}
\usage{
group_stats(df, measurement_variable, ...)
}
\arguments{
\item{df}{A dataframe containing repeat measurement data}

\item{measurement_variable}{The variable to calculate the standard deviation for}

\item{...}{Variables to group by}
}
\value{
A list containing the grouped data frame with standard deviation, and
the pooled standard deviation
}
\description{
Calculate statistical variation within groups
}
\section{Pooled standard deviation}{


This function includes the pooled standard deviation calculations described in
\href{https://nvlpubs.nist.gov/nistpubs/ir/2019/NIST.IR.6969-2019.pdf}{NISTIR 6969}
Selected Laboratory and Measurement Practices and Procedures to Support
Basic Mass Calibrations
(page 211, section 8.4).
}

\examples{

head(data_cnv)

group_stats(df = data_cnv, measurement_variable = dosage, sample, gene)

}
