% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool-sd.R
\name{pool_sd}
\alias{pool_sd}
\title{Calculated pooled standard deviation}
\usage{
pool_sd(sum_squares, sum_degrees_freedom)
}
\arguments{
\item{sum_squares}{Sum of the sum of squares across all groups}

\item{sum_degrees_freedom}{Sum of the degrees of freedom across all groups}
}
\value{
The pooled standard deviation
}
\description{
Calculated pooled standard deviation
}
