% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum-squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculate the sum of squares}
\usage{
sum_squares(degrees_freedom, stdev)
}
\arguments{
\item{degrees_freedom}{The degrees of freedom within a group}

\item{stdev}{The standard deviation within a group}
}
\value{
The squared standard deviation multiplied by degrees of freedom
}
\description{
Calculate the sum of squares
}
