% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cnv}
\alias{data_cnv}
\title{Copy number variant repeat data}
\format{
A dataframe in tidy format with columns for sample,
replicate, gene and dosage. Dosages for 3 genes are provided for each sample
replicate.
}
\source{
Dosage results are randomly generated with R.
}
\usage{
data_cnv
}
\description{
Copy number variant repeat data
}
\keyword{datasets}
