% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-sd.R
\name{group_sd}
\alias{group_sd}
\title{Calculate standard deviation with grouped variables}
\usage{
group_sd(df, measurement_variable, ...)
}
\arguments{
\item{df}{A dataframe containing repeat measurement data}

\item{measurement_variable}{The variable to calculate the standard deviation for}

\item{...}{Variables to group by}
}
\value{
A list containing the grouped data frame with standard deviation, and
the pooled standard deviation
}
\description{
Calculate standard deviation with grouped variables
}
\section{Calculation source}{


This function performs the pooled standard deviation calculations described in
\href{https://nvlpubs.nist.gov/nistpubs/ir/2019/NIST.IR.6969-2019.pdf}{NISTIR 6969}
Selected Laboratory and Measurement Practices and Procedures to Support
Basic Mass Calibrations
(page 211, section 8.4).
}

\examples{

test_df <- data.frame(
"sample" = rep(x = c(1, 2, 3), each = 3*3),
"replicate" = rep(seq(1,6,by = 1), each = 3, times = 3),
"gene" = rep(x = c("gene1", "gene2", "gene3"), times = 3*6),
"dosage" = runif(n = 54, min = 1.8, max = 2.2))

group_sd(df = test_df, measurement_variable = dosage, sample, gene)

}
